Rename BN-based abstraction helper scripts
