+++ src/dbneval_transforms.py
@@ -4,7 +4,7 @@ import os
 __thisdir = os.path.dirname(os.path.abspath(__file__))
 sys.path.insert (0, os.path.join (__thisdir))
 
-from dbnc_abstract import *
+from dbnabstr import *
 from tensorflow.keras.preprocessing.image import ImageDataGenerator
 
 # ---
